1. moved files and output utils into separate dir 2. renamed runners
